Bug 2105992: Move new function to utils folder
