+++ controllers/utils/batches_test.go
@@ -1,20 +1,4 @@
-/*
-Copyright 2022.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package controllers
+package utils
 
 import (
 	"testing"
@@ -121,12 +105,12 @@ func TestBatchTimeout(t *testing.T) {
 
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {
-			actual := calculateBatchTimeout(
+			actual := CalculateBatchTimeout(
 				tc.inputs.timeoutMinutes,
-				tc.inputs.currentBatchStartTime,
-				tc.inputs.cguStartTime,
 				tc.inputs.numBatches,
-				tc.inputs.currentBatch)
+				tc.inputs.currentBatch,
+				tc.inputs.currentBatchStartTime,
+				tc.inputs.cguStartTime)
 			assert.Equal(t, tc.expected, actual, "The expected and actual timeout should be the same.")
 		})
 	}
